move message and response into separate files
