+++ Cargo.toml
@@ -13,7 +13,7 @@ documentation = "https://docs.expo.io/versions/latest/guides/push-notifications"
 repository = "https://github.com/expo/expo-server-sdk-rust"
 
 [dependencies]
-reqwest = {version = "0.11.13", features = ["json"]}
+reqwest = {version = "0.11.13", features = ["json", "gzip"]}
 serde_json = "1"
 serde = {version = "1", features = ["derive"]}
 failure = "0.1"
